additional logic added compass and some warning messages

+++ ServiceLibrary/Models/Driver.cs
@@ -4,7 +4,7 @@ namespace ServiceLibrary.Models
 {
     public class Driver
     {
-        public double Tiredness { get; set; }
+        public int Tiredness { get; set; }
         public Hunger Hunger { get; set; }
     }
 }
